+++ packages/registry-explorer/components/Map/index.ts
@@ -1,0 +1,12 @@
+export { default as WorldMap } from "./WorldMap"
+export { default as MapLegend } from "./MapLegend"
+export { default as LegendContainer } from "./LegendContainer"
+export { default as LegendItem } from "./LegendItem"
+export { default as MapTooltip } from "./MapTooltip"
+export { default as MapSearch } from "./MapSearch"
+export { default as Sidebar } from "./Sidebar"
+export { default as CoverageSummary } from "./CoverageSummary"
+export { default as CertificateTimeline } from "./CertificateTimeline"
+export { default as CertificateDetails } from "./CertificateDetails"
+export { default as CertificateSummary } from "./CertificateSummary"
+export { default as RegistryDiff } from "./RegistryDiff"
